backend/pyproject.toml
[tool.poetry]
name = "pyproject-toml"
version = "0.1.0"
description = ""
authors = ["rajithkumar012 <rajithkumar012@gmail.com>"]
readme = "README.md"
packages = [{include = "pyproject"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.4.0"
numpy = "^2.2.3"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
